# docs/swagger.yaml
basePath: /api/v1/
definitions:
  v1.createPartInput:
    properties:
      description:
        type: string
      full_name:
        type: string
      id:
        type: string
      name:
        type: string
      targetID:
        type: string
    required:
    - description
    - full_name
    - id
    - name
    type: object
  v1.grantRoleInput:
    properties:
      role:
        type: string
      username:
        type: string
    required:
    - role
    - username
    type: object
  v1.response:
    properties:
      message:
        type: string
    type: object
  v1.signInInput:
    properties:
      name:
        type: string
      password:
        type: string
    required:
    - name
    - password
    type: object
  v1.signUpInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  v1.tokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  v1.treatiseCreateInput:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    required:
    - date
    - description
    - id
    - title
    type: object
  v1.treatiseDeleteInput:
    properties:
      title:
        type: string
    required:
    - title
    type: object
host: localhost:8080
info:
  contact: {}
  description: Hub for specified topics
  title: Hub
  version: "0.001"
paths:
  /admins/content:
    post:
      consumes:
      - application/json
      description: CreateTreatise
      parameters:
      - description: treatise info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.treatiseCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: admin CreateTreatise
      tags:
      - content
  /admins/content/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteTreatise
      parameters:
      - description: treatise info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.treatiseDeleteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: admin DeleteTreatise
      tags:
      - content
  /admins/content/{id}/parts:
    post:
      consumes:
      - application/json
      description: createPart
      parameters:
      - description: part info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.createPartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: admin createPart
      tags:
      - parts
  /admins/roles/grant-role:
    post:
      consumes:
      - application/json
      description: admin-grantrole
      parameters:
      - description: role granting info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.grantRoleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: admin GrantRole
      tags:
      - roles
  /users/sign-in:
    post:
      consumes:
      - application/json
      description: user sign in
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: user SignIn
      tags:
      - user-auth
  /users/sign-up:
    post:
      consumes:
      - application/json
      description: create user account
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: User SignUp
      tags:
      - user-auth
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  StudentsAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"